service_data/requirements.txt
celery==5.2.1
django==3.2.9
psycopg2-binary==2.9.2
flower==1.0.0
requests==2.27.1
djangorestframework==3.13.1
python-dotenv==0.19.2
redis==4.1.1
remote-pdb==2.1.0
django_filter==2.4.0